+++ test/integration/PaymentCancelTest.php
@@ -28,6 +28,7 @@ use UnzerSDK\Constants\CancelReasonCodes;
 use UnzerSDK\Resources\PaymentTypes\Invoice;
 use UnzerSDK\Resources\PaymentTypes\InvoiceSecured;
 use UnzerSDK\test\BaseIntegrationTest;
+use UnzerSDK\test\Helper\TestEnvironmentService;
 
 class PaymentCancelTest extends BaseIntegrationTest
 {
@@ -349,7 +350,7 @@ class PaymentCancelTest extends BaseIntegrationTest
      */
     public function fullCancelOnInitialInvoiceCharge($amount): void
     {
-        $this->useNon3dsKey();
+        $this->getUnzerObject()->setKey(TestEnvironmentService::getLegacyTestPrivateKey());
         /** @var Invoice $invoice */
         $invoice = $this->unzer->createPaymentType(new Invoice());
         $charge = $invoice->charge(100.0, 'EUR', self::RETURN_URL);
@@ -370,7 +371,7 @@ class PaymentCancelTest extends BaseIntegrationTest
      */
     public function partCancelOnInitialInvoiceChargeShouldBePossible(): void
     {
-        $this->useNon3dsKey();
+        $this->getUnzerObject()->setKey(TestEnvironmentService::getLegacyTestPrivateKey());
         /** @var Invoice $invoice */
         $invoice = $this->unzer->createPaymentType(new Invoice());
         $charge = $invoice->charge(100.0, 'EUR', self::RETURN_URL);
@@ -390,6 +391,7 @@ class PaymentCancelTest extends BaseIntegrationTest
      */
     public function partCancelOnInitialInvoiceSecuredChargeShouldCancelMaxUnpaidAmount(): void
     {
+        $this->getUnzerObject()->setKey(TestEnvironmentService::getLegacyTestPrivateKey());
         /** @var InvoiceSecured $invoiceSecured */
         $invoiceSecured = $this->unzer->createPaymentType(new InvoiceSecured());
 
@@ -426,6 +428,7 @@ class PaymentCancelTest extends BaseIntegrationTest
      */
     public function fullCancelOnPaidInvoiceSecuredPaymentShouldBePossible(): void
     {
+        $this->getUnzerObject()->setKey(TestEnvironmentService::getLegacyTestPrivateKey());
         /** @var InvoiceSecured $invoiceSecured */
         $invoiceSecured = $this->unzer->createPaymentType(new InvoiceSecured());